# Conflicts:
#	app/src/main/res/navigation/nav_graph.xml

+++ app/src/main/java/com/example/quick_cheque/model/Cheque.kt
@@ -1,8 +1,10 @@
 package com.example.quick_cheque.model
 
+import java.math.BigDecimal
+
 data class Cheque(
     var title: String,
     var owner: User,
-    var sumOfCheque: Int = 0,
+    var sumOfCheque: BigDecimal = BigDecimal(0),
     var membersCheque: MutableList<User> = mutableListOf(),
 )
